% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{analogsea.ext-package}
\alias{analogsea.ext}
\alias{analogsea.ext-package}
\title{Extensions to work with the analogsea package. Currently only to use ggmap::geocode at DigitalOcean}
\description{
Extensions to work with the analogsea package. Currently only to use ggmap::geocode at DigitalOcean.
}

